refactor AI files + sanitization
