+++ app/letter-generator/components/error-message.tsx
@@ -1,5 +1,5 @@
-import { Button } from '@/components/ui/button';
 import { AlertCircle, RefreshCw } from 'lucide-react';
+import { Button } from '@/components/ui/button';
 
 interface ErrorMessageProps {
   message: string;
