+++ lib/prompts/constants.ts
@@ -1,0 +1,81 @@
+// Define sensitive terms and their replacements
+export const SENSITIVE_TERMS = [
+  { term: 'revenge porn', replacement: 'non-consensual intimate content' },
+  { term: 'function normally', replacement: 'specific impact description' },
+  { term: 'criminal justice', replacement: 'legal system or legal justice' },
+  { term: 'suffering', replacement: 'specific impact description' },
+  { term: 'suffer', replacement: 'specific impact description' },
+  { term: 'victims', replacement: 'survivors' },
+  { term: 'prostitution', replacement: 'sex work' },
+  { term: 'child pornography', replacement: 'child sexual abuse material' },
+  { term: 'honour-based crimes', replacement: 'so called honour-based crimes' },
+  { term: 'mental health', replacement: 'wellbeing' },
+  { term: 'trauma', replacement: 'impact' },
+  { term: 'depression', replacement: 'emotional impact' },
+  { term: 'anxiety', replacement: 'distress' },
+  { term: 'suicide', replacement: 'severe emotional distress' },
+  { term: 'self-harm', replacement: 'personal harm' },
+];
+
+export const QUALITY_CHECK_CRITERIA = {
+  MAJOR: {
+    SENSITIVE_TERMS,
+    HALLUCINATION_TERMS: [
+      'as I mentioned earlier',
+      'as stated in my previous correspondence',
+      'as per our conversation',
+      'you have requested',
+      'you have asked me to',
+      'as you know',
+      'as we discussed',
+      'in your email',
+      'in your message',
+      'as indicated in your report',
+      'you have',
+      'as I mentioned',
+      'per our discussion',
+      'as previously noted',
+      'as explained before',
+      'in my previous email',
+      'as requested',
+      'your previous message',
+    ],
+    EVIDENCE_TERMS: [
+      'government(-|\\s)issued id',
+      'passport\\b',
+      "driver('s|\\s)licen[sc]e",
+      'national id',
+      'identity card',
+      'state(-|\\s)issued',
+      'official id',
+      'id verification process',
+      'identity verification document',
+      'verify.*identity.*document',
+      'proof of residence',
+      'proof of address',
+      'utility bill',
+      'notari[sz]ed',
+      'apostille',
+      'certified document',
+      'legal document',
+      'official documentation',
+      'government document',
+      'identification document',
+      'verification process',
+      'proof of identity',
+    ].map((term) => new RegExp(term, 'i')),
+  },
+  MINOR: {
+    STYLE_PATTERNS: [
+      /\b(please|kindly)\b/gi,
+      /\b(demand|insist|require)\b/gi,
+      /\b(immediately|urgently|asap)\b/gi,
+      /\b(must|shall|will)\b/gi,
+      /\b(legal action|lawsuit|sue)\b/gi,
+      /\b(failure to comply|failure to act)\b/gi,
+      /\b(emotional distress|mental anguish)\b/gi,
+      /\b(traumatic|devastating|horrific)\b/gi,
+      /\b(expect|anticipate|await)\b.*\b(immediate|prompt|quick)\b/gi,
+    ],
+  },
+};
